feat: refactor filehandle to new storage manager and move opencount to handle.
