--- src/new_storage/handle/mod.rs
@@ -1,6 +1,0 @@
-//! File handle module
-
-#[allow(clippy::module_inception)]
-pub mod handle;
-pub mod reader;
-pub mod writer;
