+++ src/new_storage/mod.rs
@@ -8,7 +8,6 @@ mod backend;
 mod block;
 mod block_slice;
 mod error;
-mod handle;
 mod memory_cache;
 pub mod policy;
 mod storage_manager;
@@ -21,8 +20,8 @@ pub use block::Block;
 pub use block::BLOCK_SIZE;
 pub use block_slice::{offset_to_slice, BlockSlice};
 pub use error::{StorageError, StorageResult};
-pub use handle::handle::{FileHandle, Handles, OpenFlag};
 pub use memory_cache::{CacheKey, MemoryCache};
+pub use storage_manager::handle::{FileHandle, Handles, OpenFlag};
 pub use storage_manager::StorageManager;
 pub use storage_trait::Storage;
 pub use utils::{format_file_path, format_path};
